DatabaseFundamental/Lab1_EcommerceAnalytics/requirements.txt
psycopg2-binary==2.9.9
redis==5.0.1
pymongo==4.6.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
docker==7.0.0

